applied some style and change the structure of the pages
