+++ src/components/ListPage.js
@@ -1,0 +1,50 @@
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import Card from "@material-ui/core/Card";
+import CardActions from "@material-ui/core/CardActions";
+import CardContent from "@material-ui/core/CardContent";
+import Button from "@material-ui/core/Button";
+import Typography from "@material-ui/core/Typography";
+
+const useStyles = makeStyles({
+  root: {
+    minWidth: 275,
+  },
+  bullet: {
+    display: "inline-block",
+    margin: "0 2px",
+    transform: "scale(0.8)",
+  },
+  title: {
+    fontSize: 14,
+  },
+  pos: {
+    marginBottom: 12,
+  },
+});
+
+export const ListPage = (props) => {
+  const classes = useStyles();
+  return (
+    <div>
+      <Card className={classes.root} variant="outlined">
+        <CardContent>
+          <Typography variant="h5" component="h2">
+            {props.title}
+          </Typography>
+          <Typography className={classes.pos} color="textSecondary">
+            {props.createdAt}
+          </Typography>
+          <Typography variant="body2" component="p">
+            Note: {props.description}
+          </Typography>
+        </CardContent>
+        <CardActions>
+          <Button size="small" onClick={props.handleRemove}>
+            Check detail
+          </Button>
+        </CardActions>
+      </Card>
+    </div>
+  );
+};
